API_test/requirements.txt
awscli==1.20.56
botocore==1.21.56
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.3
cycler==0.11.0
docutils==0.15.2
emoji==1.6.1
Flask==2.0.1
fonttools==4.28.2
fsspec==2021.11.0
google-api-core==2.2.1
google-api-python-client==2.28.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
mysql-connector-python==8.0.27
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.1
